+++ src/TrashLib/Radarr/CustomFormat/Guide/IRadarrGuideService.cs
@@ -1,9 +1,8 @@
 using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace TrashLib.Radarr.CustomFormat.Guide;
 
 public interface IRadarrGuideService
 {
-    Task<IEnumerable<string>> GetCustomFormatJsonAsync();
+    IEnumerable<string> GetCustomFormatJson();
 }