+++ src/VersionControl/IGitRepository.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace VersionControl;
+
+public interface IGitRepository : IDisposable
+{
+    void ForceCheckout(string branch);
+    void Fetch(string remote = "origin");
+    void ResetHard(string toBranch);
+}